fix: fix korok's child class's glob
